requirements.txt
ipdb
numpy
psutil

python-ptrace
angr
lief

# frida-tools
# r2pipe
# angr-utils